pyproject.toml
[project]
name = "theleague"
version = "0.1.0"
description = "A Year Long Sludge of a Fantasy Sports League"
readme = "README.md"
authors = [
    { name = "Jared Zirkes", email = "jaredzrks0@gmail.com" }
]
requires-python = ">=3.10.0"
dependencies = [
    "bs4>=0.0.2",
    "fsspec>=2025.5.1",
    "gcsfs>=2025.5.1",
    "lxml>=5.4.0",
    "multimodal-communication==0.1.2",
    "pandas>=2.3.0",
    "polars>=1.31.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "selenium>=4.34.2",
]

[project.scripts]
theleague = "theleague:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
